大幅にソースコードの整理を行った。
